requirements.txt
# Data Handling and Manipulation
pandas==2.2.2
numpy==1.26.0

# Machine Learning
scikit-learn==1.5.0 
# xgboost>=1.7.0
# lightgbm>=3.3.2
# catboost>=1.0.6

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Data Preprocessing and Transformation
# datetime>=4.4
# category_encoders>=2.4.0 # For categorical feature encoding

# Logging and Debugging
# logging==0.4.9.6

# For testing
pytest>=7.1.0

# Version Controling and experiment tracking 
dvc==3.51.2
dvc-gdrive==3.0.1

# Additional Useful Libraries
# requests>=2.31.0
# dotenv>=0.21.0
pickle==5.0
# joblib>=1.2.0 
argparse==1.4.0
simple-namespace==1.0.0
colorama==0.4.6
PyYAML==6.0.1

